# IELTSLISA_ZaloApp_User/IELTSLISA_ZaloApp_User/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "IELTSLISA_ZaloApp_User, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/VoucherGift/AddVoucherGift": {
      "post": {
        "tags": [
          "VoucherGift"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "giftId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "quantity",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/VoucherGift/UpdateVoucherGift": {
      "put": {
        "tags": [
          "VoucherGift"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "giftId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "quantity",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/VoucherGift/GetRandomGift": {
      "get": {
        "tags": [
          "VoucherGift"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/VoucherGift/GetVoucherGift": {
      "get": {
        "tags": [
          "VoucherGift"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}